fix up the bot.events into cogs instead to make main.py more neat
