+++ events/events_template.txt
@@ -1,0 +1,10 @@
+from discord.ext import commands
+from packages.utils import Embed
+class Events(commands.Cog):
+    def __init__(self, client):
+        self.client = client
+    @commands.Cog.listener('function_name')
+    async def event():
+        return
+def setup(client):
+    client.add_cog(Events(client))
