+++ pyproject.toml
@@ -33,6 +33,7 @@ coverage = "^5.5"
 pathos = "^0.2.7"
 aiomultiprocess = "^0.9.0"
 keras = "^2.4.3"
+sklearn = "^0.0"
 
 [tool.poetry.dev-dependencies]
